Time: 2 ms (96.84%), Space: 45 MB (7.85%) - LeetHub

+++ 0637-average-of-levels-in-binary-tree/0637-average-of-levels-in-binary-tree.java
@@ -35,7 +35,7 @@ class Solution {
                     q.offer(cur.right);
                 }
             }
-            list.add((double) sum / size);
+            list.add(sum / size);
         }
         return list;
     }
